+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## v1.1.0
+Released on: 2022-08-16
+### Added
+- Direct message protection.
+- Message format configuration.
+
 ## v1.0.0
 Released on: 2022-08-04
 ### Added